fix show post
