+++ src/pages/posts/[postId].tsx
@@ -57,7 +57,7 @@ export default Index;
 
 export const getServerSideProps: GetServerSideProps<Props> = async (ctx) => {
   const { postId } = ctx.query;
-  const post: Post = await postPageResolver(postId as string);
+  const post = await postPageResolver(postId as string);
   return {
     props: { post },
   };
