+++ src/resolvers/postPageResolver.ts
@@ -1,31 +1,12 @@
-import { Post, PostPageResponse } from 'dao/generated/graphql';
-import GraphqlPostToFrontendPost from 'logics/GraphqlPostToFrontendPost';
-import fetchApi from 'utils/fetchApi';
+import { getSdk, Post } from 'dao/generated/graphql';
+import WordpressPostToPosts from 'logics/WordpressPostToPosts';
+import createGraphqlClient from 'utils/createGraphqlClient';
 
 const postPageResolver = async (id: string): Promise<Post> => {
-  const query = `
-    query MyQuery($id: ID!) {
-      post(id: $id) {
-        content
-        id
-        postId
-        title
-        featuredImage {
-          node {
-            mediaItemUrl
-          }
-        }
-        date
-        categories {
-          nodes {
-            name
-          }
-        }
-      }
-    }`;
   const variables = { id };
-  const result: PostPageResponse = await fetchApi(query, { variables });
-  return GraphqlPostToFrontendPost(result);
+  const graphqlSdk = getSdk(createGraphqlClient());
+  const response = await graphqlSdk.postPage(variables);
+  return WordpressPostToPosts(response.post);
 };
 
 export default postPageResolver;
